backend/requirements.txt
Flask==3.1.3
Flask-Cors==6.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.44
PyMySQL==1.1.2
bcrypt==4.2.1
python-dotenv==1.1.1
requests==2.32.5
gunicorn==23.0.0
cryptography==44.0.3
PyJWT==2.9.0
